+++ Neuro/Patches/EmergencyMinigame_Patches.cs
@@ -1,0 +1,13 @@
+﻿using HarmonyLib;
+using Reactor.Utilities;
+
+namespace Neuro.Patches;
+
+[HarmonyPatch(typeof(EmergencyMinigame), nameof(EmergencyMinigame.CallMeeting))]
+public static class EmergencyMinigame_CallMeeting
+{
+    public static void Postfix(EmergencyMinigame __instance)
+    {
+        PluginSingleton<NeuroPlugin>.Instance.MeetingBegin();
+    }
+}
