+++ Neuro/Patches/PlayerPhysics_Patches.cs
@@ -1,0 +1,14 @@
+﻿using HarmonyLib;
+using Reactor.Utilities;
+using UnityEngine;
+
+namespace Neuro.Patches;
+
+[HarmonyPatch(typeof(PlayerPhysics), nameof(PlayerPhysics.SetNormalizedVelocity))]
+public static class PlayerPhysics_SetNormalizedVelocity
+{
+    public static bool Prefix(PlayerPhysics __instance, ref Vector2 direction)
+    {
+        return PluginSingleton<NeuroPlugin>.Instance.MovePlayer(ref direction);
+    }
+}
