+++ Neuro/Utils/LastSeenPlayer.cs
@@ -1,0 +1,26 @@
+﻿using System;
+
+namespace Neuro.Utils;
+
+[Serializable]
+public class LastSeenPlayer
+{
+    public string location;
+    public float time;
+    public bool dead;
+    public PlayerControl[] witnesses;
+    public float gameTimeVisible; // Total time that we've been able to see this player for
+    public float roundTimeVisible; // Total time this round that we've been able to see this player for
+    public bool sawVent; // If we've caught this player venting in front of us before
+
+    public LastSeenPlayer(string location, float time, bool dead)
+    {
+        this.location = location;
+        this.time = time;
+        this.dead = dead;
+        this.gameTimeVisible = 0f;
+        this.roundTimeVisible = 0f;
+        this.sawVent = false;
+        this.witnesses = null;
+    }
+}
